wip: Recompose macro structure & Fix naming
