+++ src/components/index.ts
@@ -1,10 +1,7 @@
 export { Layout } from "./layout/layout.jsx";
 export { Column } from "./column/column.jsx";
-export { Task } from "./task/task.jsx";
 export { Menu } from "./menu/menu.jsx";
 export { Builder } from "./builder/builder.jsx";
 export { Editor } from "./editor/editor.jsx";
 export { Export } from "./export/export.jsx";
-export { Wallet } from "./wallet/wallet.jsx";
 export { Sidebar } from "./sidebar/sidebar.jsx";
-export { Dao } from "./dao/dao";
