+++ src/entities/index.ts
@@ -1,0 +1,3 @@
+export { Task } from "./task";
+export { Token } from "./token";
+export { Wallet } from "./wallet";
