+++ src/entities/task/index.ts
@@ -1,0 +1,5 @@
+import { TaskCard } from "./ui/card";
+
+export const Task = {
+    Card: TaskCard,
+};
