+++ src/entities/task/ui/card.scss
