+++ src/entities/token/ui/fungible.scss
