+++ src/entities/token/ui/fungible.tsx
