+++ src/entities/wallet/index.ts
@@ -1,0 +1,8 @@
+import { WalletComponent } from "./ui";
+import { useSelector, WalletSelectorContextProvider } from "./ui/providers";
+
+export const Wallet = {
+    Selector: WalletComponent,
+    SelectorContextProvider: WalletSelectorContextProvider,
+    useSelector,
+};
