+++ src/entities/wallet/ui/index.scss
