+++ src/entities/wallet/ui/providers.tsx
@@ -1,5 +1,11 @@
-import React, { useCallback, useEffect, useState } from "react";
-import { map, distinctUntilChanged } from "rxjs";
+// @ts-ignore
+import nearWalletIconUrl from "@near-wallet-selector/near-wallet/assets/near-wallet-icon.png";
+// @ts-ignore
+import senderIconUrl from "@near-wallet-selector/sender/assets/sender-icon.png";
+// @ts-ignore
+import myNearWalletIconUrl from "@near-wallet-selector/my-near-wallet/assets/my-near-wallet-icon.png";
+// @ts-ignore
+import mathWalletIconUrl from "@near-wallet-selector/math-wallet/assets/math-wallet-icon.png";
 import { setupWalletSelector } from "@near-wallet-selector/core";
 import type { WalletSelector, AccountState } from "@near-wallet-selector/core";
 import { setupModal } from "@near-wallet-selector/modal-ui";
@@ -8,24 +14,8 @@ import { setupNearWallet } from "@near-wallet-selector/near-wallet";
 import { setupMyNearWallet } from "@near-wallet-selector/my-near-wallet";
 import { setupSender } from "@near-wallet-selector/sender";
 import { setupMathWallet } from "@near-wallet-selector/math-wallet";
-
-// import wallet icons
-const nearWalletIconUrl = new URL(
-    "../../node_modules/@near-wallet-selector/near-wallet/assets/near-wallet-icon.png",
-    import.meta.url
-);
-const senderIconUrl = new URL(
-    "../../node_modules/@near-wallet-selector/sender/assets/sender-icon.png",
-    import.meta.url
-);
-const myNearWalletIconUrl = new URL(
-    "../../node_modules/@near-wallet-selector/my-near-wallet/assets/my-near-wallet-icon.png",
-    import.meta.url
-);
-const mathWalletIconUrl = new URL(
-    "../../node_modules/@near-wallet-selector/math-wallet/assets/math-wallet-icon.png",
-    import.meta.url
-);
+import React, { useCallback, useEffect, useState } from "react";
+import { map, distinctUntilChanged } from "rxjs";
 
 declare global {
     interface Window {
@@ -51,17 +41,20 @@ export const WalletSelectorContextProvider: React.FC<any> = ({ children }) => {
     const init = useCallback(async () => {
         const _selector = await setupWalletSelector({
             network: window.NEAR_ENV,
+
             modules: [
-                setupNearWallet({ iconUrl: nearWalletIconUrl.href }),
-                setupSender({ iconUrl: senderIconUrl.href }),
-                setupMyNearWallet({ iconUrl: myNearWalletIconUrl.href }),
-                setupMathWallet({ iconUrl: mathWalletIconUrl.href }),
+                setupNearWallet({ iconUrl: nearWalletIconUrl }),
+                setupSender({ iconUrl: senderIconUrl }),
+                setupMyNearWallet({ iconUrl: myNearWalletIconUrl }),
+                setupMathWallet({ iconUrl: mathWalletIconUrl }),
             ],
         });
+
         const _modal = setupModal(_selector, {
             contractId: window.nearConfig.MULTICALL_FACTORY_ADDRESS,
             theme: "dark",
         });
+
         const state = _selector.store.getState();
 
         setAccounts(state.accounts);
@@ -119,7 +112,7 @@ export const WalletSelectorContextProvider: React.FC<any> = ({ children }) => {
     );
 };
 
-export function useWalletSelector() {
+export function useSelector() {
     if (!WalletSelectorContext) {
         throw new Error("useWalletSelector must be used within a WalletSelectorContextProvider");
     }
