+++ src/pages/dao/funds/index.scss
@@ -3,7 +3,7 @@
 @use "sass/mixin";
 @use "sass/size";
 
-.funds-tab {
+.FundsTab {
     display: grid;
     grid-template: "balances" 1fr / 1fr;
